Refactor schema management: consolidate entity loading in EntityManager
