--- sql/queries/check-installation.sql
@@ -1,3 +1,0 @@
-SELECT COUNT(*) as count 
-FROM sqlite_master 
-WHERE type = ? AND name = ?
