--- sql/queries/get-version.sql
@@ -1,3 +1,0 @@
-SELECT value 
-FROM system_settings 
-WHERE key = ?
