--- sql/queries/load-entity-definition.sql
@@ -1,3 +1,0 @@
-SELECT * 
-FROM system_entities 
-WHERE name = ? AND is_active = ?
