--- sql/queries/load-entity-fields.sql
@@ -1,4 +1,0 @@
-SELECT * 
-FROM system_fields 
-WHERE entity_id = ? AND is_active = ? 
-ORDER BY order_index ASC
